dataset/python-mutated/test_zip_reader.py
import datetime
import io
import struct
import zipfile
from unittest.mock import MagicMock, patch
import pytest
from source_s3.v4.zip_reader import DecompressedStream, RemoteFileInsideArchive, ZipContentReader, ZipFileHandler

@pytest.fixture
def mock_s3_client():
    if False:
        for i in range(10):
            print('nop')
    return MagicMock()

@pytest.fixture
def mock_config():
    if False:
        return 10
    return MagicMock(bucket='test-bucket')

@pytest.fixture
def zip_file_handler(mock_s3_client, mock_config):
    if False:
        print('Hello World!')
    return ZipFileHandler(mock_s3_client, mock_config)

def test_fetch_data_from_s3(zip_file_handler):
    if False:
        return 10
    zip_file_handler._fetch_data_from_s3('test_file', 0, 10)
    zip_file_handler.s3_client.get_object.assert_called_with(Bucket='test-bucket', Key='test_file', Range='bytes=0-9')

def test_find_signature(zip_file_handler):
    if False:
        i = 10
        return i + 15
    zip_file_handler.s3_client.head_object.return_value = {'ContentLength': 1024}
    zip_file_handler._fetch_data_from_s3 = MagicMock(return_value=b'test' + ZipFileHandler.EOCD_SIGNATURE + b'data')
    result = zip_file_handler._find_signature('test_file', ZipFileHandler.EOCD_SIGNATURE)
    assert ZipFileHandler.EOCD_SIGNATURE in result

def test_get_central_directory_start(zip_file_handler):
    if False:
        while True:
            i = 10
    zip_file_handler._find_signature = MagicMock(return_value=b'\x00' * 16 + struct.pack('<L', 12345))
    zip_file_handler._find_signature.return_value = b'\x00' * 16 + struct.pack('<L', 12345)
    assert zip_file_handler._get_central_directory_start('test_file') == 12345

def test_get_zip_files(zip_file_handler):
    if False:
        i = 10
        return i + 15
    zip_file_handler._get_central_directory_start = MagicMock(return_value=0)
    zip_file_handler._fetch_data_from_s3 = MagicMock(return_value=b'dummy_data')
    with patch('io.BytesIO', return_value=MagicMock(spec=io.BytesIO)):
        with patch('zipfile.ZipFile', return_value=MagicMock(spec=zipfile.ZipFile)):
            (result, cd_start) = zip_file_handler.get_zip_files('test_file')
            assert cd_start == 0

def test_decompressed_stream_seek():
    if False:
        return 10
    mock_file = MagicMock(spec=io.IOBase)
    mock_file.read = MagicMock()
    mock_file.read.return_value = b'test'
    mock_file.tell.return_value = 0
    file_info = RemoteFileInsideArchive(uri='test_file.csv', last_modified=datetime.datetime(2022, 12, 28), start_offset=0, compressed_size=100, uncompressed_size=200, compression_method=zipfile.ZIP_STORED)
    stream = DecompressedStream(mock_file, file_info)
    assert stream.seek(2) == 2

def test_decompressed_stream_seek_out_of_bounds():
    if False:
        while True:
            i = 10
    mock_file = MagicMock(spec=io.IOBase)
    mock_file.read = MagicMock()
    mock_file.read.return_value = b'test'
    mock_file.tell.return_value = 0
    file_info = RemoteFileInsideArchive(uri='test_file.csv', last_modified=datetime.datetime(2022, 12, 28), start_offset=0, compressed_size=4, uncompressed_size=8, compression_method=zipfile.ZIP_STORED)
    stream = DecompressedStream(mock_file, file_info)
    assert stream.seek(10) == 8

def test_zip_content_reader_readline():
    if False:
        i = 10
        return i + 15
    mock_stream = MagicMock(spec=DecompressedStream)
    mock_stream.read.return_value = b'test\n'
    reader = ZipContentReader(mock_stream, encoding='utf-8')
    assert reader.readline() == 'test\n'

def test_zip_content_reader_read():
    if False:
        return 10
    mock_stream = MagicMock(spec=DecompressedStream)
    mock_stream.read.return_value = b'test_data'
    reader = ZipContentReader(mock_stream, encoding='utf-8')
    assert reader.read(4) == 'test'

def test_zip_content_reader_readline_newline_combinations():
    if False:
        return 10
    mock_stream = MagicMock(spec=DecompressedStream)
    mock_stream.read.side_effect = [b'test1\n', b'']
    reader = ZipContentReader(mock_stream, encoding='utf-8')
    assert reader.readline() == 'test1\n'
    mock_stream.read.side_effect = [b'test2\r', b'']
    reader = ZipContentReader(mock_stream, encoding='utf-8')
    assert reader.readline() == 'test2\r'
    mock_stream.read.side_effect = [b'test3\r', b'\n', b'']
    reader = ZipContentReader(mock_stream, encoding='utf-8')
    assert reader.readline() == 'test3\r\n'

def test_zip_content_reader_iteration():
    if False:
        i = 10
        return i + 15
    mock_stream = MagicMock(spec=DecompressedStream)
    mock_stream.read.side_effect = [b'line1\n', b'line2\r', b'line3\r\n', b'line4\n', b'']
    reader = ZipContentReader(mock_stream, encoding='utf-8')
    lines = list(reader)
    assert lines == ['line1\n', 'line2\r', 'line3\r\n', 'line4\n']